refactor: divide productFilter constants into different components
