+++ src/components/Filters/BrandFilter/index.js
@@ -1,0 +1,2 @@
+import BrandFilter from "./BrandFilter";
+export default BrandFilter;
