+++ src/components/Filters/CategoriesFilter/index.js
@@ -1,0 +1,2 @@
+import CategoriesFilter from "./CategoriesFilter";
+export default CategoriesFilter;
