+++ src/components/Filters/ColorFilter/index.js
@@ -1,0 +1,2 @@
+import ColorFilter from "./ColorFilter";
+export default ColorFilter;
