+++ src/components/Filters/DiscountFilter/index.js
@@ -1,0 +1,2 @@
+import DiscountFilter from "./DiscountFilter";
+export default DiscountFilter;
