+++ src/components/Filters/PriceFilter/index.js
@@ -1,0 +1,2 @@
+import PriceFilter from "./PriceFilter";
+export default PriceFilter;
