+++ src/components/Filters/RaitingFilter/index.js
@@ -1,0 +1,2 @@
+import RaitingFilter from "./RaitingFilter";
+export default RaitingFilter;
